create separate files for data classes / Unit test fix
