+++ data/src/main/java/com/bottlerocketstudios/brarchitecture/data/model/BranchDto.kt
@@ -1,0 +1,10 @@
+package com.bottlerocketstudios.brarchitecture.data.model
+
+import android.os.Parcelable
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+import kotlinx.parcelize.Parcelize
+
+@JsonClass(generateAdapter = true)
+@Parcelize
+data class Branch(@Json(name = "name") val name: String?) : Parcelable
