+++ data/src/test/java/com/bottlerocketstudios/brarchitecture/data/model/CommitTest.kt
@@ -7,11 +7,12 @@ import org.junit.Test
 class CommitTest : BaseTest() {
     @Test
     fun commit_defaultFields_whenDefaultConstructor() {
-        val commit = Commit(null, null, null, null, null)
+        val commit = Commit(null, null, null, null, null, null)
         assertThat(commit.parents).isNull()
         assertThat(commit.date).isNull()
         assertThat(commit.message).isNull()
         assertThat(commit.type).isNull()
         assertThat(commit.hash).isNull()
+        assertThat(commit.links).isNull()
     }
 }
